+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 1.0.0
 * Fix replication agent creation failure #1
+* Unsuccessful action should now raise RubyAem::Error, replacing failure result
 * Move resource classes to RubyAem::Resources module
 
 ### 0.9.0